backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
pyyaml==6.0.1
httpx==0.27.0
pytest==8.2.0
python-dotenv==1.0.1
openai==1.25.0
pytest-asyncio==0.23.6 